Split modal component, put delete function out of the component
