+++ src/Components/Shared/Modal/ModalMessage/modalMessage.module.css
@@ -1,0 +1,43 @@
+.overlay {
+    position: fixed;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    top: 0;
+    right: 0;
+    width: 100%;
+    height: 100vh;
+    background-color: #000000BF;
+  }
+  
+  .container {
+    min-width: 300px;
+    width: 40%;
+    max-width: 600px;
+    background-color: #FFFFFF;
+  }
+  
+  .header, .headerSuccess, .content {
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    box-sizing: border-box;
+    width: 100%;
+    padding: 10px;
+    font-size: 20px;
+  }
+  
+  .header {
+    background-color: #ff0000;
+    color: #FFFFFF;
+  }
+  
+  .headerSuccess {
+    background-color: #008000;
+    color: #FFFFFF;
+  }
+  
+  .modalCloseButton:hover {
+    cursor: pointer;
+  }
+
