+++ src/Components/TimeSheets/Form/index.js
@@ -1,6 +1,7 @@
 import { useEffect, useState } from 'react';
 import { useParams } from 'react-router-dom';
-import { ModalConfirm, ModalMessage } from '../../Shared/Modal/Modal';
+import ModalConfirm from '../../Shared/Modal/ModalConfirm/index';
+import ModalMessage from '../../Shared/Modal/ModalMessage/index';
 import Input from '../../Shared/Inputs';
 import Datepicker from '../../Shared/Datepicker';
 import Select from '../../Shared/Select/index';
